send user activation email
