+++ statina/API/external/api/api_v1/endpoints/update.py
@@ -21,6 +21,46 @@ router = APIRouter()
 LOG = logging.getLogger(__name__)
 
 
+@router.get("/validate_user")
+async def validate_user(
+    username: str,
+    verification_hex: str,
+    adapter: StatinaAdapter = Depends(get_nipt_adapter),
+):
+    update_user: User = find.user(user_name=username, adapter=adapter)
+    response = RedirectResponse("/batches")
+    if not update_user:
+        response.set_cookie(key="info_type", value="danger")
+        response.set_cookie(
+            key="user_info",
+            value="No such user in the database",
+        )
+    elif update_user.verification_hex == verification_hex and update_user.role == "unconfirmed":
+        try:
+            update_user.role = "inactive"
+            update.update_user(adapter=adapter, user=update_user)
+            email_form = FormDataRequest(
+                sender_prefix=email_settings.sender_prefix,
+                request_uri=email_settings.mail_uri,
+                recipients=email_settings.admin_email,
+                mail_title="New user request",
+                mail_body=f"User {update_user.username} ({update_user.email}) requested new account <br>"
+                f'Follow <a href="{email_settings.website_uri}">link</a> to activate user',
+            )
+            email_form.submit()
+            response.set_cookie(
+                key="user_info",
+                value="Email confirmed! Your account will be activated after manual review",
+            )
+        except Exception as e:
+            response.set_cookie(key="info_type", value="danger")
+            response.set_cookie(
+                key="user_info",
+                value=f"Backend error ({e.__class__.__name__})! Your account will be activated after manual review",
+            )
+    return response
+
+
 @router.post("/update_user")
 async def update_user(
     request: Request,
@@ -31,8 +71,20 @@ async def update_user(
         return RedirectResponse(request.headers.get("referer"))
     form = await request.form()
     update_user: User = find.user(email=form["user_email"], adapter=adapter)
-    update_user.role = form["role"]
+    old_role = update_user.role
+    new_role = form["role"]
+    update_user.role = new_role
     update.update_user(adapter=adapter, user=update_user)
+    if old_role in ["inactive", "unconfirmed"] and new_role not in ["inactive", "unconfirmed"]:
+        email_form = FormDataRequest(
+            sender_prefix=email_settings.sender_prefix,
+            request_uri=email_settings.mail_uri,
+            recipients=update_user.email,
+            mail_title="Your account has been activated",
+            mail_body=f'Your <a href="{email_settings.website_uri}">statina</a> account ({update_user.username})'
+            f" has been activated!",
+        )
+        email_form.submit()
     return RedirectResponse(request.headers.get("referer"))
